Fixes some bugs and completes the further testing
The GUI bug which caused it to either not display the correct amount of homes or not display the homes' names correctly has been fixed!
The plugin works 90% perfectly. There is still one known bug to which a solution is in order. This bug allows the /delhome command to "delete" anything from the config file, which messes up with the GUI