Finished Plugin: Project Completed
This is the PalaceHomes final version (1.0). All bugs and issues have now been fixed and the plugin is ready to use. I still advise against tampering with the config.yml file, because the storage methods and codes are slightly complex and you might break them down if you delete/change something you shouldn't.